retrieve uncached form when form state not yet cached.
This change attempts to load a form with drupal_retrieve_form when a cached form is not found by form_get_cache.  This was causing a small error when a user backed up (i.e. with the browser's back button) onto a form whose build state was no longer cached.  instead of giving the user annoying javascript messages, we reload the form using the current form state.

A short screencast exposing a minor bug:
http://youtu.be/Oj4CZu8Xo_8